initial sidebar redo
